Separated old API and updated ColorMap API.
